reorg functionality to diff files
